Check for Secrets in Kubernetes
This script uses the Kubernetes Python client library to list all secrets in all namespaces in the cluster. If there are no secrets present, it will print a message indicating that there are no secrets in the namespace. If secrets are present, it will print a message indicating that there are secrets in the namespace.

It is important to note that this script only checks for the presence of secrets, and it does not check the security of the secrets themselves. It is recommended to use best practices for securing secrets in Kubernetes, such as using encrypted secrets, using a secrets management solution, and controlling access to secrets using Role-Based Access Control (RBAC).