Configfinder
Here's a Python script that can search for configuration files within a Kubernetes cluster. This script uses the os library to search for files in the directory /etc/kubernetes. It then checks if each file ends with either .yaml or .yml. This assumes that you are using these common file types/extensions. If a configuration file is found, it is added to a list and then the list is sent to console.