containerprivs
This Python script that can check for container privileges using the docker library. So that's a good one to use. It uses the docker library to connect to the Docker API and retrieve information about the specified container. The check_container_privileges function takes a container_id as input and returns a Boolean indicating whether the container is running with privileged mode enabled. Then, you guessed it, the script prints a message to the console indicating whether the container is running with privileged mode enabled.