Create basic container security
This is a script that checks host machines using the os library to run shell commands that check for three things.  The check_security function checks if the firewall (ufw) is enabled, if SELinux is enabled, and if the system is up-to-date. If any of these checks fail, a message is added to a list of security issues. 
If the host machines are using a different OS then you need to make updates accordingly.